Fix game editor trying to edit OOB tiles
Fixes a bug causing the editor to sometimes edit tiles at
coordinates that are out of the game board's boundaries.

This would occur due to the game camera's `get_subject_tile_by_coordinates`
function processing coordinates that are equal to the camera's subject's
dimensions.

As a result, if the mouse was exactly on the edge of the game board's
tiles and the user clicked, the tile to edit's coordinates would
be just one tile out of bounds.